+++ noodles-vcf/src/header/sample.rs
@@ -1,14 +1,12 @@
 //! VCF header sample record.
 
-pub mod key;
-
-pub use self::key::Key;
-
 use std::{error, fmt};
 
 use indexmap::IndexMap;
 
 use super::{record, Record};
+
+const ID: &str = "ID";
 
 /// A VCF header sample record (`SAMPLE`).
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -65,7 +63,7 @@ impl fmt::Display for Sample {
         f.write_str(record::Key::Sample.as_ref())?;
         f.write_str("=<")?;
 
-        write!(f, "{}={}", Key::Id, self.id())?;
+        write!(f, "{}={}", ID, self.id())?;
 
         for (key, value) in &self.fields {
             write!(f, r#",{}={}"#, key, value)?;
@@ -83,7 +81,7 @@ pub enum TryFromRecordError {
     /// The record is invalid.
     InvalidRecord,
     /// A required field is missing.
-    MissingField(Key),
+    MissingField(&'static str),
 }
 
 impl error::Error for TryFromRecordError {}
@@ -113,10 +111,10 @@ fn parse_struct(fields: Vec<(String, String)>) -> Result<Sample, TryFromRecordEr
 
     let id = it
         .next()
-        .ok_or(TryFromRecordError::MissingField(Key::Id))
-        .and_then(|(k, v)| match k.parse() {
-            Ok(Key::Id) => Ok(v),
-            _ => Err(TryFromRecordError::MissingField(Key::Id)),
+        .ok_or(TryFromRecordError::MissingField(ID))
+        .and_then(|(k, v)| match k.as_ref() {
+            ID => Ok(v),
+            _ => Err(TryFromRecordError::MissingField(ID)),
         })?;
 
     let fields = it.collect();